% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allca}
\alias{allca}
\title{All cascades from Vosoughi et al}
\format{A data frame with 126301 rows and 7 variables}
\usage{
data(allca)
}
\description{
A dataset containing a selection of properties of the cascades analysed in Vosoughi et al 2018.
}
\details{
\itemize{
  \item date. data of apparition of the cascade
  \item rumor_id. Id of the rumor associated to the cascade
  \item breadth. breadth of the cascade (\emph{ie} max number of parallel retweets) 
  \item size. size of the cascade (\emph{ie} number of unique retweets) 
  \item depth. depth of the cascade (\emph{ie} max number successive of unique retweets)
  \item veracity. veracity of the associated rumor (in  \{"TRUE","FALSE","MIXED"\})
  \item time_last. number of second between the apparition of the cascade and the time of the last retweet
}
}
\keyword{datasets}
