% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDrafts.R
\name{cascades}
\alias{cascades}
\title{Cascade number of retweet and length of cascad of differents tweets}
\usage{
cascades(N = 200, K = 10, M = 5, betadistrib = runif(200, 0, 1),
  jdistrib = runif(200, 0, 1), utility = seq(0, 1, length.out = 5),
  repetition = 2, seed = 100, tl = 500)
}
\arguments{
\item{N}{: number of agents}

\item{K}{: max number of neighbours (not used in this version)}

\item{M}{: number of messages}

\item{betadistrib}{: vector os size N with distribution of beta for each agents}

\item{utility}{: vector of size M, the distribution of intrinsic utility for each messages}

\item{repetition}{: number of time that the agents will check their timeline}

\item{seed}{: initial number of messages}

\item{tl}{: maximum messages that agents can acess to}

\item{seed}{: initial maximum number of tweeet that the `seeder` (first tweeter) can do}
}
\value{
a list with information on each messages
}
\description{
Cascade number of retweet and length of cascad of differents tweets
}
