% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_analyse.R
\name{cascades3D.list}
\alias{cascades3D.list}
\title{Wrapper for Cascades3D that automatically convert a list of arguments}
\usage{
cascades3D.list(listofparameters, metrics = c("size", "depth",
  "breadth"))
}
\arguments{
\item{listofparameters}{a list of parameters normally given to cascades3D}
}
\value{
same as cascades3D
}
\description{
Wrapper for Cascades3D that automatically convert a list of arguments
}
