% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeTools.R
\name{ccfd}
\alias{ccfd}
\title{Calculate the CCFD following \emph{Vosoughi et al. (2018)}}
\usage{
ccfd(size)
}
\arguments{
\item{size}{: sample of sizes}
}
\value{
a order table with two column with first column : frequency and second column the probality of the frequency
}
\description{
Calculate the CCFD following \emph{Vosoughi et al. (2018)}
}
