% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_analyse.R
\name{checkIfNumberGood}
\alias{checkIfNumberGood}
\title{Take a folder name and return the list of id of the folders that have exact same number of parameters and scores}
\usage{
checkIfNumberGood(folder, inv = T)
}
\arguments{
\item{folder}{the name of a folder as the one used in getlistfoldsubfold}
}
\description{
Take a folder name and return the list of id of the folders that have exact same number of parameters and scores
}
