% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{depths}
\alias{depths}
\title{depths: 
the recursive function called to compute the depth}
\usage{
depths(cascade, root)
}
\arguments{
\item{cascade:}{a edgelist that represent a cascade}

\item{root:}{the id of the root node}
}
\value{
list of depth of each branch
}
\description{
depths: 
the recursive function called to compute the depth
}
