% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{diff}
\alias{diff}
\title{Stupid Diff Function
distance between two sample a and b}
\usage{
diff(a, b)
}
\arguments{
\item{a}{and b: sample of any size}
}
\value{
a list of dist
}
\description{
Stupid Diff Function
distance between two sample a and b
}
