% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{euclidediff}
\alias{euclidediff}
\title{Binned Frequencies distance
distance between the frequencies of a and b, the frequencies beings calculates for logarithmics categories}
\usage{
euclidediff(a, b, nbin = 20)
}
\arguments{
\item{a}{and b: sample of any size}
}
\value{
a list of dist
}
\description{
Binned Frequencies distance
distance between the frequencies of a and b, the frequencies beings calculates for logarithmics categories
}
