% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{falseca}
\alias{falseca}
\title{False Cascades from Vosoughi et al}
\usage{
data(falseca)
}
\description{
Only the false cascades from in Vosoughi et al 2018.
}
\seealso{
\code{\link{allca}}
}
\keyword{datasets}
