% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{filter}
\alias{filter}
\title{simple function to handle parameter space made of multiple dimension parameter}
\usage{
filter(vec, filter)
}
\arguments{
\item{vec}{a vector potentially multidimensional}

\item{filter}{the index of the vector that we want to keep}
}
\description{
simple function to handle parameter space made of multiple dimension parameter
}
