% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphviz.R
\name{formatEdgeList}
\alias{formatEdgeList}
\title{transfrom a cascade made of numeric index into character
if not igraph will fill missing vertices with fake one}
\usage{
formatEdgeList(cascade, i = 0)
}
\arguments{
\item{cascade}{a cascade}

\item{i}{an integer to identify the cascade, 0 if nothing is given}
}
\value{
an edgelist formated to be read by `graph_from_edgelist`
}
\description{
transfrom a cascade made of numeric index into character
if not igraph will fill missing vertices with fake one
}
