% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_analyse.R
\name{fromTestScoresVSParam}
\alias{fromTestScoresVSParam}
\title{Function that helps to run new test to check if score correspond do parameters.
allow to be sure that parameters and scores read from the folders make sense.}
\usage{
fromTestScoresVSParam(theparam, thescores, numberparam = 20,
  repet = 10, modelwrapper = cascades3D.list, data = allru$size,
  scorefun = quantilediff, rumor = F)
}
\arguments{
\item{numberparam}{number of parameter that will be tested}

\item{rep}{number of time the parameter set will be repeted}
}
\description{
Function that helps to run new test to check if score correspond do parameters.
allow to be sure that parameters and scores read from the folders make sense.
}
