% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{generalPartition}
\alias{generalPartition}
\title{Generate a table  n subvector of random size}
\usage{
generalPartition(n, d)
}
\arguments{
\item{n}{number of row}

\item{d}{number of column}
}
\value{
a table of dim nxd
}
\description{
Generate a table  n subvector of random size
}
