% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{generateCascades}
\alias{generateCascades}
\title{generateCascades initialize a list of cascades of retweet with only the first tweet}
\usage{
generateCascades(nc, seeders, rumours, time)
}
\arguments{
\item{nc}{: number of cascades}

\item{seeders}{: list or number of possible agents that will start the cascade}

\item{time}{: the time at wich start the cascades}

\item{random}{: list or number of possible rumors where to choose the cascades}
}
\value{
a list of size nc with the id of the first agents and the rumor
}
\description{
generateCascades initialize a list of cascades of retweet with only the first tweet
}
