% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{generateDistribeFromFrequencies}
\alias{generateDistribeFromFrequencies}
\title{Generate a Distribution of n value splited in   reparted among all class of the vector `cls` following the frequenceies `freq`}
\usage{
generateDistribeFromFrequencies(cls, n, freq)
}
\arguments{
\item{cls}{all possible classe}

\item{n}{total number of value}

\item{freq}{final frequencies for each class}
}
\value{
a vector of size `n` with value from `cls` that respect the frequencies in `freq`
}
\description{
Generate a Distribution of n value splited in   reparted among all class of the vector `cls` following the frequenceies `freq`
}
