% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_analyse.R
\name{getAllparameters}
\alias{getAllparameters}
\title{From a main folder get all parameters
In theory, the folder `mainfold` should contain the results of an abc as done by ./abcdir/abc.R
which, in theory again, should look like a set of subfolder `mainfoldresults/mainfoldresultsX` where X depends on the number of simulation and the way they are divided in abc.R}
\usage{
getAllparameters(mainfold, log = F, lim = NULL, check = T)
}
\arguments{
\item{mainfold}{mainfolder with allscores}

\item{log}{if true, show the name of the folder scanned}

\item{lim}{if lim is a vector of int then take the folder that have index in lim, if lim is a character, directly take lim as the list of folders}

\item{check}{if check is true then when check if all parameters led to good simulation (ie number of parameters -== number of score)
note that with check use of checkIfNumberGood should be useless}
}
\value{
a dataframe with all parameters
}
\description{
From a main folder get all parameters
In theory, the folder `mainfold` should contain the results of an abc as done by ./abcdir/abc.R
which, in theory again, should look like a set of subfolder `mainfoldresults/mainfoldresultsX` where X depends on the number of simulation and the way they are divided in abc.R
}
