% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_analyse.R
\name{getAllscores}
\alias{getAllscores}
\title{From a main folder get all scores 
In theory, the folder `mainfold` should contain the results of an abc as done by ./abcdir/abc.R
which, in theory again, should look like a set of subfolder `mainfoldresults/mainfoldresultsX` where X depends on the number of simulation and the way they are divided in abc.R}
\usage{
getAllscores(mainfold, idscores = c("kc", "kokc", "qc", "kcrev",
  "kokcrev"), metrics = c("depth", "breadth", "size"), log = T,
  lim = NULL)
}
\arguments{
\item{mainfold}{mainfolder with allscores}

\item{idscores}{the id of the score}

\item{metrics}{the ids of the metrics use for the score}

\item{log}{if true, show the name of the folder scanned}
}
\value{
a list of list of list
}
\description{
From a main folder get all scores 
In theory, the folder `mainfold` should contain the results of an abc as done by ./abcdir/abc.R
which, in theory again, should look like a set of subfolder `mainfoldresults/mainfoldresultsX` where X depends on the number of simulation and the way they are divided in abc.R
}
