% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_analyse.R
\name{getbest}
\alias{getbest}
\title{Return the best set of parameter given a score an}
\usage{
getbest(allscores, allparameters.dataframe, s, m)
}
\arguments{
\item{allscores}{a list of list of score}

\item{allparameters.dataframe}{a dataframe with all parameters}

\item{s}{id of the function use to compute the score}

\item{m}{metric use to compute the score}
}
\value{
the list of parameter that gave the best score
}
\description{
Return the best set of parameter given a score an
}
