% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getDepth}
\alias{getDepth}
\title{getDepth: 
Gives the depth of a tree}
\usage{
getDepth(cascade)
}
\arguments{
\item{cascade:}{a edgelist that represent a cascade}
}
\value{
integer
}
\description{
getDepth: 
Gives the depth of a tree
}
