% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getMaxBreadth}
\alias{getMaxBreadth}
\title{getMaxBreadth: 
Gives the max breadth of a tree}
\usage{
getMaxBreadth(cascade)
}
\arguments{
\item{cascade:}{a edgelist that represent a cascade}
}
\value{
integer
}
\description{
getMaxBreadth: 
Gives the max breadth of a tree
}
