% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getRumors}
\alias{getRumors}
\title{add the size of the cascade of the same rumors togethers}
\usage{
getRumors(tablres, metric = "size")
}
\description{
add the size of the cascade of the same rumors togethers
}
