% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_analyse.R
\name{getSimuGivenIndices}
\alias{getSimuGivenIndices}
\title{go within a folder with result and extract the good simulation}
\usage{
getSimuGivenIndices(ind, expefold, full = F)
}
\arguments{
\item{ind}{integer of the indice (ie id) of the simulation}

\item{expefold}{folder within which we have to look}

\item{full}{boolean saying is the full simulation need to be retrieves}
}
\value{
a list with the score and the list of parameters
}
\description{
go within a folder with result and extract the good simulation
}
