% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getSummaryCascade}
\alias{getSummaryCascade}
\title{getSummaryCascade: from a list of cascade, return a list with the size, depth and max breadth of each cascades}
\usage{
getSummaryCascade(cascade)
}
\arguments{
\item{cascade}{: a list of cascades}
}
\value{
a  table with three column ("size","depth","breadth") and a number of row equal to the length of the input list
}
\description{
getSummaryCascade: from a list of cascade, return a list with the size, depth and max breadth of each cascades
}
