% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeTools.R
\name{getUCols}
\alias{getUCols}
\title{return a vector of color that follow the values of utilities}
\usage{
getUCols(utilities)
}
\arguments{
\item{utilites}{: a vector with value we want to match}
}
\value{
a vector of length = `length(utilities)` of colors with names=sort(names(utilities))
}
\description{
return a vector of color that follow the values of utilities
}
