% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{kldiff}
\alias{kldiff}
\title{More kindof KL
distance between two samples a and b by looking at the difference between the log o all percentile of two sample}
\usage{
kldiff(a, b)
}
\arguments{
\item{a}{and b: sample of any size}
}
\value{
a number
}
\description{
More kindof KL
distance between two samples a and b by looking at the difference between the log o all percentile of two sample
}
