% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mixedca}
\alias{mixedca}
\title{Mixed Cascades from Vosoughi et al}
\usage{
data(mixedca)
}
\description{
Only the mixed cascades from in Vosoughi et al 2018.
}
\seealso{
\code{\link{allca}}
}
\keyword{datasets}
