% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeTools.R
\name{plot2dens}
\alias{plot2dens}
\title{plot posteriors distribution against priors}
\usage{
plot2dens(A = NULL, B = NULL, C = NULL, from = NULL, to = NULL,
  prior = NULL, cols = c(alpha("red", 0.8), alpha("blue", 0.8),
  alpha("yellow", 0.8)), hdr = F, yaxt = NULL, ...)
}
\arguments{
\item{A}{: a vector with posterior}

\item{B}{: a vector with posterior}

\item{prior}{: a vector with posterior}
}
\description{
plot posteriors distribution against priors
}
