% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeTools.R
\name{plotCCFD}
\alias{plotCCFD}
\title{plot the CCFD following \emph{Vosoughi et al. (2018)} on a log-log axis}
\usage{
plotCCFD(s, pch = 20, cex = 1, xlab = "", ...)
}
\arguments{
\item{s}{: sample of sizes}

\item{...}{: parameters passed to plot}
}
\value{
nothing
}
\description{
plot the CCFD following \emph{Vosoughi et al. (2018)} on a log-log axis
}
