% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_analyse.R
\name{plotPosteriorsCheck}
\alias{plotPosteriorsCheck}
\title{Same before but print distribution function that helps to run test to check if score correspond do parameters.
and then}
\usage{
plotPosteriorsCheck(theposterior, thescores, samplepost = 100,
  repet = 10, modelwrapper = cascades3D.list, data = allru$size,
  scorefun = quantilediff, rumor = F, type = "all", clrs = "score",
  alberto = F, ...)
}
\arguments{
\item{clrs}{if "score" the curves are colored givent the score function used}

\item{...}{parameters transmitted to plot}

\item{numberparam}{number of parameter that will be tested}

\item{rep}{number of time the parameter set will be repeted}
}
\description{
Same before but print distribution function that helps to run test to check if score correspond do parameters.
and then
}
