% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeTools.R
\name{pointsCCFD}
\alias{pointsCCFD}
\title{add points of the CCFD following \emph{Vosoughi et al. (2018)}}
\usage{
pointsCCFD(s, pch = 20, cex = 1, ...)
}
\arguments{
\item{s}{: sample of sizes}

\item{...}{: parameters passed to plot}
}
\value{
nothing
}
\description{
add points of the CCFD following \emph{Vosoughi et al. (2018)}
}
