% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeTools.R
\name{pointsCCFD_U}
\alias{pointsCCFD_U}
\title{plot the CCFD following \emph{Vosoughi et al. (2018)} on a log-log axis and color it given utility}
\usage{
pointsCCFD_U(s, pch = 20, cex = 2, col = NULL, lenged = TRUE, ...)
}
\arguments{
\item{s}{: a two dimension vector with size of cascade and utility of the cascade for a simulation}

\item{...}{: parameters passed to plot}
}
\value{
nothing
}
\description{
plot the CCFD following \emph{Vosoughi et al. (2018)} on a log-log axis and color it given utility
}
