% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{probaSelection}
\alias{probaSelection}
\title{probaSelection a return the probability for a list of tweet of given  utility to be retweeted given a value of beta}
\usage{
probaSelection(u, beta)
}
\arguments{
\item{u}{: vector of utility of tweet}

\item{beta}{: value of beta}
}
\value{
a vector of probability of the size length(u)
}
\description{
probaSelection a return the probability for a list of tweet of given  utility to be retweeted given a value of beta
}
