% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{quantilediff}
\alias{quantilediff}
\title{Quantil Diff Function
distance between two samples a and b by looking at the difference between the log o all percentile of two sample}
\usage{
quantilediff(a, b, log = T, probs = seq(0, 1, 0.01))
}
\arguments{
\item{a}{and b: sample of any size}
}
\value{
a number
}
\description{
Quantil Diff Function
distance between two samples a and b by looking at the difference between the log o all percentile of two sample
}
