% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_analyse.R
\name{randomCascades.list}
\alias{randomCascades.list}
\title{Wrapper for random cascade}
\usage{
randomCascades.list(listofparameters, ...)
}
\arguments{
\item{listofparameters}{a list of parameters normally given to cascades3D}
}
\value{
same as randomCascades
}
\description{
Wrapper for random cascade
}
