% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{randomCascades}
\alias{randomCascades}
\title{randomCascades: A slightly (almost not) modified version of the original model by Alex & Damien
Original model is in tools/FNM.R}
\usage{
randomCascades(Nmin = 1e+05, Nmax = 50000, t_steps = 50, y_max = 1,
  runs = 1, mu = 1e-05, tau = 40, alpha = 0, conformity = F)
}
\value{
dataframe that allow to compare easily to the result of the model of the cascade
}
\description{
randomCascades: A slightly (almost not) modified version of the original model by Alex & Damien
Original model is in tools/FNM.R
}
