% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{removeDuplicate}
\alias{removeDuplicate}
\title{removeDuplicate return index of the vector v without the replicate and by randomly selecting amoung the replicate}
\usage{
removeDuplicate(v)
}
\arguments{
\item{v}{: a vector}
}
\value{
a vector of indice of v
}
\description{
removeDuplicate return index of the vector v without the replicate and by randomly selecting amoung the replicate
}
