% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_analyse.R
\name{reruns}
\alias{reruns}
\title{Same before but just return list of the simulation rerun}
\usage{
reruns(theposterior, thescores = NULL, samplepost = 100, repet = 10,
  modelwrapper = cascades3D.list, data = allru$size,
  scorefun = quantilediff, rumor = F, type = "all", log = F, ...)
}
\arguments{
\item{numberparam}{number of parameter that will be tested}

\item{rep}{number of time the parameter set will be repeted}
}
\description{
Same before but just return list of the simulation rerun
}
