% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDrafts.R
\name{simplemodel}
\alias{simplemodel}
\title{Simplemodel with only one message being retweeted}
\usage{
simplemodel(agents = 2:200, K = 10, betadistrib = runif(0, 1),
  jdistrib = runif(0, 1), utility = runif(1), repetition = 10,
  M = 5, seed)
}
\arguments{
\item{K}{: max number of neighbours}

\item{utility}{: intrinsic utility of the message}

\item{repetition}{: number of time that the agents will check their timeline}

\item{M}{: number of messageG}

\item{N}{: number of agents}

\item{betarange}{: range for parameter beta}
}
\description{
Simplemodel with only one message being retweeted
}
