% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tomat}
\alias{tomat}
\title{tomat from a data frame create a matrix}
\usage{
tomat(dt, fun = mean, measure = "size")
}
\arguments{
\item{dt}{: dataframe}

\item{fun}{: a function returning a single number from a sample}

\item{measure}{: the measurement on which will be applied the function fun}
}
\value{
a matrix
}
\description{
tomat from a data frame create a matrix
}
