% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trueca}
\alias{trueca}
\title{True Cascades from Vosoughi et al}
\usage{
data(trueca)
}
\description{
Only the true cascades from in Vosoughi et al 2018.
}
\seealso{
\code{\link{allca}}
}
\keyword{datasets}
